requirements (1).txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# MongoDB Driver
motor==3.3.2
pymongo==4.6.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6

# CORS Support
python-dotenv==1.0.0

# Date/Time Handling
python-dateutil==2.8.2

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0

# Code Quality (Optional)
black==23.12.1
flake8==7.0.0
mypy==1.8.0